pyproject.toml
[tool.poetry]
name = "social-signals"
version = "0.1.0"
description = "A package for harvesting social signal feeds."
authors = ["Olivier Dupuis <odupuis@lantrns.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.2"
numpy = {version = "^1.26.0", python=">=3.9,<3.13"}
requests = "^2.31.0"
google-cloud-bigquery = "^3.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
    "ignore:Deprecated call to `pkg_resources.declare_namespace`:DeprecationWarning",
]
